% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplefunction.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{add_two_numbers}
\usage{
add_two_numbers(input1, input2)
}
\arguments{
\item{input1}{Numeric value to form part of the sum}

\item{input2}{Numeric value to form part of the sum}
}
\value{
The sum of the input values
}
\description{
A function to add together two input numbers and return their sum
}
\details{
Given two numeric inputs, adds them together and returns the sum value.
}
\examples{
x <- add_two_numbers(3.5, 4)

}
