% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inheritParameters.R
\name{bar}
\alias{bar}
\title{bar two things}
\usage{
bar(a_thing, another_thing)
}
\arguments{
\item{a_thing}{This is the first argument}

\item{another_thing}{This is the second argument}
}
\description{
bar two things
}
