% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creativePlot.R
\name{creativePlot}
\alias{creativePlot}
\title{creativePlot}
\usage{
creativePlot(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{a string name for variable x in dataframe df}

\item{vary}{a string name for variable y in dataframe df}
}
\value{
A plot  of \code{vary} versus \code{varx}
}
\description{
This function produces a ggplot of \code{vary} versus \code{varx}
}
\examples{
df.test <- data.frame(x = rnorm(10),
                      y = rnorm(10))
creativePlot(df.test, 'x', 'y')

}
