% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Title}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{vary}{}
}
\value{
plot
}
\description{
Title
}
\examples{
df.test <- data.frame(x=rnorm(10),y=rnorm(10)) diamond_plot(df.test)
}
