% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{Example function}
\usage{
example_function(x, y)
}
\arguments{
\item{x}{one vector}

\item{y}{another vector}
}
\value{
This function returns the sum of the two vectors as another vector
}
\description{
This is a an example function
}
\details{
It takes two vectors and adds them together
}
\examples{
x <- 1
y <- 1
example_function(x = x, y = y)
}
\seealso{
\link[base]{sum}
}
\author{
Matt Simmons mattsimmons@email.com
}
