\name{examplePackage-package}
\alias{examplePackage-package}
\alias{examplePackage}
\docType{package}
\title{\packageTitle{examplePackage}}
\description{\packageDescription{examplePackage}}
\section{Package Content}{\packageIndices{examplePackage}}
\author{\packageAuthor{examplePackage}}
\section{Maintainer}{\packageMaintainer{examplePackage}}
\keyword{package}
