% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePackage.R
\docType{package}
\name{examplePackage}
\alias{examplePackage}
\alias{examplePackage-package}
\title{Infrastructure Tools for Organizing New Projects}
\description{
This package collects various scripts and templates
we've found to be useful in organizing projects.
}
