% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_diamonds.R
\name{explore_diamonds}
\alias{explore_diamonds}
\title{explore_diamonds
this function prints a nice markdowm summary of the diamonds data set}
\usage{
explore_diamonds()
}
\description{
explore_diamonds
this function prints a nice markdowm summary of the diamonds data set
}
\examples{
explore_diamonds()
}
\keyword{explore}
