% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleFunctions.R
\name{factors}
\alias{factors}
\title{Calculate the factors of an integer}
\usage{
factors(int)
}
\arguments{
\item{int}{A positve integer.}
}
\value{
A vector containing all factors if a positive integer. If a negative integer is provided, a message will be returned.
}
\description{
Returns the a vector containg all factors of a positive integer.
}
\note{
This function doesn't check to make sure that the value of 'int' is an integer.
}
\examples{
# Returing the factors of an integer
factors(1283)
# Storing the factors of an integer in a variable
facts <- factors(1653787)
facts
# Message returned when trying to factor a negative integer
factors(-1283)
# integer(0) is returned when trying to factor a decimal
factors(1283.1)
}
\keyword{factor}
\keyword{factors}
