% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map data by state}
\usage{
fars_map_state(state.num, year, datadir = NULL)
}
\arguments{
\item{state.num}{Numeric of the US state to be plotted (must be between 1 and 50).}

\item{year}{Numeric of the year of data to be accessed.}

\item{datadir}{Character string of the data directory. Default is NULL, in which case
the package loads the built in datasets.}
}
\value{
Nothing. Plots map with accident locations for state and year as function side effect.
}
\description{
Plot a map with locations of accident data for given state and year.
}
\note{
State id's available are:
  1  2  4  5  6  8  9 10 11 12 13 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34
  35 36 37 38 39 40 41 42 44 45 46 47 48 49 50 51 53 54 55 56
}
\examples{
fars_map_state(1, 2013)

}
