% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load csv data}
\usage{
fars_read(filename, datadir = NULL)
}
\arguments{
\item{filename}{Character string with the filepath.}

\item{datadir}{Character string of the data directory. Default is NULL, in which case
the package loads the built in datasets.}
}
\value{
Returns a tibble dataframe of the parsed csv file.
}
\description{
A simple wrapper around readr read_csv to read and load raw csv data files.
}
\examples{
fars_read('accident_2015.csv.bz2')

}
