% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years}
\usage{
fars_read_years(years, datadir = NULL)
}
\arguments{
\item{years}{Vector of numerics of the years of data to be loaded.}

\item{datadir}{Character string of the data directory. Default is NULL, in which case
the package loads the built in datasets.}
}
\value{
List of data frames giving month and year, each corresponding to
  the dataset for the year specified in the input.
}
\description{
Load multiple datasets from a vector of the years of data to be loaded.
}
\note{
Data is assumed to be in the same folder as the script.
}
