% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize years}
\usage{
fars_summarize_years(years, datadir = NULL)
}
\arguments{
\item{years}{Numeric vector of years of data to be loaded}

\item{datadir}{Character string of the data directory. Default is NULL, in which case
the package loads the built in datasets.}
}
\value{
Data frame summary of the number of records by month, by year.
}
\description{
Load and summarize number of data records by year over several datasets
}
\examples{
fars_summarize_years(c(2013,2015))

}
