% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not-in.R
\name{\%not in\%}
\alias{\%not in\%}
\title{Not In}
\usage{
x \%not in\% table
}
\arguments{
\item{x}{a value}

\item{table}{a vector or table of values}
}
\value{
This function true if x is not in y
}
\description{
This is the opposite of in
}
\details{
It does the opposite of what \code{in} does.
}
\examples{
x <- 1
y <- c(2,3,4)
x \%not in\% y
}
\seealso{

}
\author{
Matt Simmons mattsimmons@email.com
}
