% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphizer.R
\name{graphizer}
\alias{graphizer}
\title{Graphizer}
\usage{
graphizer(n_people, n_qualities, n_quality_levels, n_rounds)
}
\arguments{
\item{n_people}{The number of vertexes. Can't be more than 26.}

\item{n_qualities}{The number of qualities to compare. Can't be more than 26.}

\item{n_quality_levels}{The number of possible levels those qualities can have}

\item{n_rounds}{The number of times to try to generate an edge by comparing two vertexes qualities}
}
\value{
This function will return an igraph graph.
}
\description{
Silly function that generates an arbitrary graph
}
\details{
Generates a list of people with 'qualities' and then creates probable edges based on how far their qualities are from each other
}
\examples{
\dontrun{
g <- graphizer(10, 10, 10, 100)
}
}
\author{
Matt Simmons mattsimmons@email.com
}
