% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-not-na.R
\name{is.not.na}
\alias{is.not.na}
\title{Is not NA}
\usage{
is.not.na(x)
}
\arguments{
\item{x}{a value or vector of values}
}
\value{
This function returns if x is not na
}
\description{
This is the opposite of is.na
}
\details{
It does the opposite of what is.na
}
\examples{
x <- NA
y <- is.not.na(x)
z <- 1
q <- is.not.na(z)
}
\seealso{
\link[base]{is.na}
}
\author{
Matt Simmons mattsimmons@email.com
}
