% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclass.R
\name{logtransformed}
\alias{logtransformed}
\title{Log-Transform a Numeric Vector}
\usage{
logtransformed(NumericVector = NULL)
}
\arguments{
\item{NumericVector}{A numeric vector you would like to log-transform}
}
\value{
A list of two objects:
  \item{InputVector}{Input numeric vector}
  \item{logTransformVec}{Input numeric vector log-transformed}
}
\description{
This is an unnecessary function I created for the purposes of
  instruction
}
\details{
This function is pretty self explanatory
}
\examples{
saveout<-logtransformed(NumericVector = c(5.21, 2.03, 1.49, 13.28,
                                        474.10, 21.81, 3.19, 1.53))
saveout$InputVector
saveout$logTransformVec
}
