% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numeric of the year of data to be included in the filename.}
}
\value{
Character string of the filename composed of base name + year
}
\description{
Construct filename for a dataset given the year of data collection
}
