% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{merge}
\alias{merge}
\title{Merges two sorted lists}
\usage{
merge(x, y)
}
\arguments{
\item{x}{a sorted list}

\item{y}{a sorted list}
}
\value{
a sorted list containing elements of both lists
}
\description{
Merges two sorted lists
}
\examples{
merge(c(1,2), c(3,4))
}
