% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{mergesort}
\alias{mergesort}
\title{Sorts a list recursively using the mergesort algorithm}
\usage{
mergesort(x)
}
\arguments{
\item{x}{A list}
}
\value{
A sorted list
}
\description{
Sorts a list recursively using the mergesort algorithm
}
\examples{
mergesort(c(1,3,2,4,2)) returns c(1,2,2,3,4)
}
