% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{my_addition}
\alias{my_addition}
\title{Add two numbers incorrectly}
\usage{
my_addition(x, y)
}
\arguments{
\item{x}{This is a number that we care about. Must be Nx1 vector.}

\item{y}{This is the other number.}
}
\value{
The sum of the two numbers, but it's incorrect. It will be a numeric object.
}
\description{
Add two numbers incorrectly
}
\examples{
my_addition(1, 2)
my_addition(3, 4)
}
