% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password_gen.R
\name{password_gen}
\alias{password_gen}
\title{Password Generator}
\usage{
password_gen(password_length = 8, special_chars = NULL)
}
\arguments{
\item{password_length}{the length in characters of the password}

\item{special_chars}{an optional string vector with allowable special characters to use, if needed}
}
\value{
This returns a random password of length password_length
}
\description{
Creates a QF-valid password
}
\details{
Creates a random password that will be accepted by various corporate systems
}
\examples{
x <- password_gen()
y <- password_gen(password_length = 3)
}
\author{
Matt Simmons mattsimmons@email.com
}
