% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diamonds.R
\name{plot_diamonds}
\alias{plot_diamonds}
\title{plot_diamonds}
\usage{
plot_diamonds(rows = 1000)
}
\arguments{
\item{rows}{and integer for sampling the diamonds data}
}
\description{
plot_diamonds
}
\examples{
plot_diamonds(999)
plot_diamonds(10001)
}
\keyword{plot}
