% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot_func}
\alias{plot_func}
\title{plot y as a function of x}
\usage{
plot_func(n, min, max, func)
}
\arguments{
\item{n}{size of x.}

\item{min}{min of x.}

\item{max}{max of x.}

\item{func}{a string formula, y as a function of x. e.g. 'y = x ^ 2'}
}
\value{
The plot of \code{y} as a function of \code{x}
}
\description{
plot y as a function of x
}
\examples{
plot_func(100, 0, 10, 'y = 10 * x')
}
