% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleFunctions.R
\name{powers}
\alias{powers}
\title{Raise a number to a specfied power}
\usage{
powers(num, exp = 2)
}
\arguments{
\item{num}{A numeric value.}

\item{exp}{A numeric value. The default value is 2.}
}
\value{
A numeric value. If no value is provided for 'num', a message will be returned.
}
\description{
Returns an numeric value, or a message if the value of 'num' is not provided.
}
\note{
This function doesn't check to make sure that the values of 'num' or 'exp' are numeric.
}
\examples{
# Calculate the square of an number
powers(3)
# Calculate the square-root of an number
powers(9,0.5)
}
\keyword{powers}
