% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primes}
\alias{primes}
\title{The First 100,008 Primes}
\format{A dataframe with 10008 rows and one column (p)}
\source{
https://primes.utm.edu/lists/small/100000.txt
}
\usage{
data(primes)
}
\description{
A dataset of the first 100008 prime numbers (from 2 to 1299827).
}
\references{
C. Caldwell. The Prime Pages. 1994-2018 The University of Tennessee at Martin.
}
\keyword{numbers}
\keyword{prime}
