% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-bitre-data.R
\name{pull_bitre_data}
\alias{pull_bitre_data}
\title{Pull BITRE data}
\usage{
pull_bitre_data(
  base_url,
  resource_id,
  limit = 5000,
  offset = 1,
 
    sort = "Year desc, Month_Num desc, Airline desc, Departing_Port asc, Arriving_Port asc",
  records_format = "objects"
)
}
\arguments{
\item{base_url}{The base data.gov.au API URL}

\item{resource_id}{The specific resource ID code from the data.gov.au website}

\item{limit}{The number of rows you want to pull}

\item{offset}{the rownumber start}

\item{sort}{a text string, in SQL style, indicating how you'd like to sort the records}

\item{records_format}{leave this as objects for the moment}
}
\value{
This function will return a data frame of the BITRE on time performance data.
}
\description{
Uses the CKAN data.gov.au API to pull records - tuned for the BITRE API.
}
\details{
Retrieves a dataframe, with some limited filtering capabilities
}
\examples{
\dontrun{
data <- pull_bitre_data <- function(base_url = url, resource_id = id)
}
}
\author{
Matt Simmons mattsimmons@email.com
}
