% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshFromOldPackages.R
\name{refreshFromOldPackages}
\alias{refreshFromOldPackages}
\title{refreshFromOldPackages}
\usage{
refreshFromOldPackages(old_file = "Current_Packages.csv")
}
\arguments{
\item{old_file}{The name of the file that will be used to install packages.}
}
\value{
This function installs all the packages from the csv.
}
\description{
Uses a csv containing package names to install packages en-masse. Use in conjunction with refreshFromOldPackages
}
\details{
Uses a csv containing package names to install packages en-masse. Use in conjunction with refreshFromOldPackages
}
\examples{
\dontrun{
write.csv(data.frame(Package = "dplyr"), "Current_Packages.csv")
refreshFromOldPackages()
}
}
\author{
Matt Simmons mattsimmons@email.com
}
