% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveCurrentPackagesData.R
\name{retrieveCurrentPackagesData}
\alias{retrieveCurrentPackagesData}
\title{retrieveCurrentPackagesData}
\usage{
retrieveCurrentPackagesData(file = "Current_Packages.csv")
}
\arguments{
\item{file}{The name of the file that the packages will be written to}
}
\value{
This function writes down a file to the file location with all currently installed packages.
}
\description{
Retrieves all currently stored packages and writes them to a file for use by refreshFromOldPackages
}
\details{
Retrieves all currently stored packages and writes them to a file for use by refreshFromOldPackages
}
\examples{
\dontrun{
retrieveCurrentPackagesData()
retrieveCurrentPackagesData(file = "another-file.csv")
}
}
\author{
Matt Simmons mattsimmons@email.com
}
