% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{shuffle}
\usage{
shuffle(data, seed = NA)
}
\arguments{
\item{data}{a data frame}

\item{seed}{a random seed to fix the shuffling - optional}
}
\value{
Returns the dataframe, shuffled.
}
\description{
shuffle a dataset
}
\details{
Takes a dataset and randomly orders the rows. Good to make sure there's no row-dependence before partitioning.
}
\examples{
df <- data.frame(x = 1:100, y = 101:200)
shuffled_df <- shuffle(df)
}
\author{
Matt Simmons mattsimmons@email.com
}
