% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_table.R
\name{standardize_table}
\alias{standardize_table}
\title{Rate comparison after adjusting for a confounder}
\usage{
standardize_table(df, response, explanatory, confounder, mix = c(0.5, 0.5))
}
\arguments{
\item{df}{A dataframe with three categorical variables (chr or factor)}

\item{response}{character vector indicating the name of the response variable}

\item{explanatory}{character vector indicating the name of the explanatory variable}

\item{confounder}{character vector indicating the name of the confounding variable}

\item{mix}{a two-entry numerical vector representing the percentages of each level of the confounder used for standardization}
}
\value{
a list with entries "unstd" for the unstandardized table, and "std" for the standardized table using the weights from \code{mix}
}
\description{
Rate comparison after adjusting for a confounder
}
\examples{
whickham<-read.table("https://s3.amazonaws.com/pbreheny-data-sets/whickham.txt",
    sep = "\t",header=TRUE)
whickham$over65 <- ifelse(whickham$Age=="65-","Yes","No")
table1<-standardize_table(
    df = whickham,
    response = "Survival",
    explanatory = "Smoking",
    confounder = "over65",
    mix = c(.7, .3)
)
table1$unstd
table1$std
}
