% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_plot.R
\name{star_plot}
\alias{star_plot}
\title{Star Plot}
\usage{
star_plot(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{a string name for variable x in the dataframe df}

\item{vary}{a string name for variable y in the dataframe df}
}
\value{
A plot with star markers of \code{vary} versus \code{varx}
}
\description{
This function produces a ggplot plot of \code{vary} versus \code{varx} with
green star markers.
}
\examples{

df_test <- data.frame(x = rnorm(10), y = rnorm(10))

diamond_plot(df_test)


}
