% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creativePlot.R
\name{creativePlot}
\alias{creativePlot}
\title{creativePlot}
\usage{
creativePlot(n)
}
\arguments{
\item{n}{the amount of stars}
}
\value{
A plot  of n random pink stars
}
\description{
This function produces a ggplot of n random pink stars
}
\examples{
df.test <- data.frame(x = rnorm(10),
                      y = rnorm(10))
creativePlot(df.test, 'x', 'y')

}
