% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Diamand Plot
This funciton produces a ggplot of \code{vary} versus \code{varx} with
diamand markers.}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{a strong name for variable x in the dataframe df}

\item{vary}{a strong name for variable y in the dataframe df}
}
\value{
A plot with diamand markers of \code{vary} versus \code{varx}
}
\description{
Diamand Plot
This funciton produces a ggplot of \code{vary} versus \code{varx} with
diamand markers.
}
\examples{

df.test <- data.frame(x = rnorm(10),
                      y = rnorm(10))

diamond_plot(df.test, 'x', 'y')

}
