% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Diamond Plot}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{a data frame}

\item{varx}{a string name for variable x in the data frame df}

\item{vary}{a string name for variable y in the data frame df}
}
\value{
A plot with diamond markers of \code{vary} verses \code{varx}
}
\description{
This function produces a ggplot of \code{vary} versus \code{varx}
}
\examples{
df.test <- data.frame(x = rnorm(10),
                      y = rnorm(10))
diamond_plot(df.test, 'x', 'y')

}
