% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExamplePackage.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Diamond Plot}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{a data frame}

\item{varx}{a string name for variable x in the data fram df}

\item{vary}{a string name for variable y in the data fram df}
}
\value{
A plot with diamond markers of \code{vary} versus \code{varx}
}
\description{
This function produces a ggplot of \code{vary} versus \code{varx}
}
\examples{
df.test <- data.fram(x=rnorm(10),
                     y=rnorm(10))

diamond_plot(df_test,'x','y')

}
