% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePackage.R
\docType{package}
\name{examplePackage}
\alias{examplePackage}
\alias{examplePackage-package}
\title{examplePackage}
\description{
An example package
}
\details{
A test to try out making packages
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
