% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting.R
\name{quicksort}
\alias{quicksort}
\title{Recursively sorts a list by selecting a random pivot}
\usage{
quicksort(x)
}
\arguments{
\item{x}{A list}
}
\value{
Returns a sorted version of `x`
}
\description{
Recursively sorts a list by selecting a random pivot
}
\examples{
quicksort(c(1,3,2,4,2)) returns c(1,2,2,3,4)
}
